LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/window-onerror-runtime-error-throw-expected.txt

CONSOLE MESSAGE: line 32: bar
CONSOLE MESSAGE: line 1401: Error: Tried to create a test with file_is_test

PASS window.onerror: runtime scripterrors 
